backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain
langchain-openai
langchain-community
youtube-transcript-api
faiss-cpu